+++ src/KardFinancial.Test/Unit/MockServer/Organizations/Placements/ListTest.cs
@@ -20,6 +20,7 @@ public class ListTest : BaseMockServerTest
                   "id": "id",
                   "attributes": {
                     "name": "name",
+                    "displayName": "displayName",
                     "organizationId": "organizationId",
                     "availableSlots": 1,
                     "contentStrategyId": "contentStrategyId"
@@ -38,6 +39,7 @@ public class ListTest : BaseMockServerTest
                   "id": "id",
                   "attributes": {
                     "name": "name",
+                    "displayName": "displayName",
                     "organizationId": "organizationId",
                     "availableSlots": 1,
                     "contentStrategyId": "contentStrategyId"