+++ src/KardFinancial.Test/Unit/MockServer/Organizations/Placements/UpdateTest.cs
@@ -30,6 +30,7 @@ public class UpdateTest : BaseMockServerTest
               "id": "id",
               "attributes": {
                 "name": "name",
+                "displayName": "displayName",
                 "organizationId": "organizationId",
                 "availableSlots": 1,
                 "contentStrategyId": "contentStrategyId"